Display images from their data instead of path
This is so that images can be synchronized across devices.
The problem is that this incurs a have performance penalty, mostly
visible when opening the settings tab or navigating out of a team form
into the team list.
This is due to the fact that on each rerender the huge image URIs have
to be reconstructed. React native docs advice against using URIs for big
images because it is unperformant:
https://facebook.github.io/react-native/docs/images#uri-data-images