Add default handicap type to all team-related forms
This is pretty important, because the forms start with a handicap type
selected, even though the value is not propogated into the redux-form
state - it needs to be set in the initialValues.
If it is not set, there will be an incomplete object in the state with
only the handicapValue but not the handicapType if the user did not
touch the slider. Since the user should assume that since the sliders
value was set to something it is saved as such, we need to have the
defaultValue safeguards.
It is also helpful for backwards-compatibility.
The change in the actions/sessions is for making double sure