Create functionless <FilterSessions /> screen
Header icons, like the one that should be on the right to click to the
FilterSessions screen on the Sessions screen, can only exist on
StackNavigators. Therefore, it was necessary to remove the
TopTabNavigator in the Sessions screen, which did not have a practical
purpose anyways. The only thing to reproduce from it is the orange
seperator between the header and the rest of the screen content and the
padding between the header title and the header seperator