Add redux-saga to the project and convert selectCourse to saga to solve race condition
Previously there was a race condition in selectCourse, when clicking on
a course that has to be fetched, navigating back and clicking on define
course. With a arbitrary delay with `setTimeout()`, this sequence of
steps would always result in the new course layout being overwritten by
the course from the server, even though `Define Layout` was pressed.

Introducing `takeLatest` solves this problem