Fix leaderboard scrolling not working
It was due to the fact that the FlatList was bigger vertically than the screen
and it rendered the list items beyond the bottom of the screen. Because
of the fact that the elements were rendered (off screen) it didn't allow
to scroll. The solution is to limit the height of the FlatList to fit
inside the screen.