Ignore errors due to missing permissions when deleting positioning information for markPropeties when deleting binding
When a user binds his device via a QR code to a mark that is not owned
by him, when deleting the binding he is expected to fail to update the
corresponding markProperties. We can ignore that error, because the
update is not necessary. Refer to the ticket comments for more information.