Fix the bug where clicking on the back button quickly several times in the course creation screen made the "Save course?" dialog not show up
This works by overriding the back button behaviour and the Android
system back button behaviour.

Also added a `once` to the header back button because I think it makes
more sense that way. It's not necessary for the fix, so it can be
removed in the future if decided that it's not needed.