Add a competitor list to the event details screen
This adds the list both to the owner and a regular competitor.

For some reason I couldn't get the "there's no competitors currently"
text to work with the ./styles file in the common file, so I had to
inline the styles. If using styles from ./styles it would just be plain
text.