Copy over newly created marks into the next race's course by name
To solve an issue where when you create a new course with a new mark you
can't access that course anymore, because the newly created mark is not
properly copied over to the next course.

This has a problem though, because you have to use the name of the mark
to get the markConfigurationId because there's no other way to find it.
However, this doesn't work well with marks that have the same name.