Only sync events joined anonymously in syncEventList
This is to avoid a bug where a device would not see the bindings on
another device when logging in because it would overwrite the checkin in
the event inventory. Here's the scenario where the bug would
occur:
1) Create Event A on device A
2) Log in into the same account on device B
3) Bind device B to a competitor in Event A
4) Sometimes (I don't know when and why) the device A will have been
logged out. If that happens, it still has the event in the inventory.
Log in to the same user on device A.
5) It will not see that device B has been bound to a competitor in the
event, because it will have overwritten the new checkIn with the
saveCheckInToEventInventory call in the syncEventList function