Fix background being blacked out when sharing link to event from event overview screen
The problem came from the ActionSheet component being in the middle of
the dismissal animation and the share menu coming up at the same time,
which created the issue for some reason because of the ActionSheet.
Adding a 1ms delay with setTimeout seems to make the animations go one
after the other and fixes the issue.