Update the positioning of a markProperties & mark when specifying its geolocation
Instead of having to wait for the course to be saved, the positioning
will be immediately updated when the user specifies the location in the
Geolocation screen or by pressing the "Ping current location" button.

Will also update the positioning of the regatta mark if the course has
already been created.