Prevent creation of multiple mark-device mappings when pressing the "Use this device" button in the TrackerBinding screen
To make the behavior of the mark binding be closer to the one that the
server provides us with when we use the `createCourse` call. The
`createCourse` call prevents multiple mark-device mappings from being
created when you repeatedly save the same course.

This prevents problems with the deletion of mark bindings. When you
delete a binding in the MarkTracking screen, it only deletes a single
binding, which is why we have to enforce this here.